prevent multiple popups on screen when 'show[ing] on click'
This is, indeed, a bug in the `react-popup` library, as suspected.  This commit is a tolerable work-around.  A bug has been reported on the library repo, and a PR has been made, but this work-around will work as well if not better in our case, and doesn't make us dependent on a new npm release.