move styling to sass
Cherry-picking this from my working branch, because working around all this mess to get a proper production deploy working is silly.

Original commit message:
---
This primary motivation for this is to get access to global style definitions in these stylesheet (colours etc.), but it also means they get compiled, bundled, and minified properly (they're currently getting inlined in the HTML).  The stylesheets are not being encapsulated properly at the component level anyway, so there is nothing to be lost doing it this way instead.