add image opacity filter
With the previous commits in place, I now think it works well to simply use a css filter to adjust the opacity of the tutorial images.  As a bonus, they can be 'brought up' on mouse-over, which I think works really nicely, personally, especially given that there are two different images which can come in and out independently.