adjust build-time environment variables
For reasons unclear, using `JEKYLL_ENV=production` causes relative URLs to be prepended with `/pages/sul-cidr/scriptchart` (presumably by the `gh-pages` gem), which, of course, breaks everything.  But we need an env variable to toggle inclusion of the Google Analytics snippet, so here is a solution.  It means the GA snippet is not included when deploying to GitHub Pages, but that's a good thing :)